Add tree parameter to URL queries
The newer versions of Cloudbees/Jenkins require that the "tree"
parameter be passed in REST API requests. This patch is a limited
fix for the jenkins library, and is only meant to address the use
cases for the noironetworks automation. A more general fix could
be that the application passes the fields that it's interested in
to the jenkins class APIs, which are then used to generate the
appropriate URL when making the query to Jenkins..